requirements.txt
flask>=2.3,<3
google-generativeai>=0.3,<0.4
python-dotenv>=1.0,<2
